+++ R/add_reviewer.R
@@ -1,22 +1,23 @@
 #' Add reviewer
 #'
-#' @param x a \code{\link{goldrake}} object
+#' @param gold a \code{\link{goldrake}} object
 #' @param name (chr) Reviewer's name
 #' @param surname  (chr, default = NULL, optional) Reviewer's surname
-#' @param inspector (lgl, default = FALSE) Is a reviewer a reviever's
+#' @param inspector (lgl, default = FALSE) Is a reviewer a reviewer's
 #'        reviewers?
 #' @param seed (num, default = 1) seed for the case sampling
 #'
-#' @return invisibly, the updated version of x
+#' @return invisibly, the updated version of gold
 #' @export
 #'
 #' @examples
 #' with_reviewer <- goldrake(mtcars) %>%
 #'     set_gold_classes() %>%
+#'     balance_groups_by(cyl, vs) %>%
 #'     add_reviewer("Reviewer 1")
 #'
 #' with_reviewer
-add_reviewer <- function(x,
+add_reviewer <- function(gold,
     name, surname = NULL, inspector = FALSE, seed = 1) {
     UseMethod("add_reviewer")
 }
@@ -24,12 +25,12 @@ add_reviewer <- function(x,
 #' @rdname add_reviewer
 #' @export
 add_reviewer.default <- function(
-    x, name, surname = NULL, inspector = FALSE, seed = 1
+    gold, name, surname = NULL, inspector = FALSE, seed = 1
 ) {
     ui_stop(paste(
-        "The {ui_field('x')} input in {ui_code('add_reviewer')}",
+        "The {ui_field('gold')} input in {ui_code('add_reviewer')}",
         "must be inherited from the {ui_value('goldrake')} class.\n",
-        "It is of class {ui_value(class(x))}.\n",
+        "It is of class {ui_value(class(gold))}.\n",
         "Please, provide an object of class {ui_value('goldrake')}."
     ))
 }
@@ -37,19 +38,32 @@ add_reviewer.default <- function(
 #' @rdname add_reviewer
 #' @export
 add_reviewer.goldrake <- function(
-    x, name, surname = NULL, inspector = FALSE, seed = 1
+    gold, name, surname = NULL, inspector = FALSE, seed = 1
 ) {
     if (length(name) != 1) ui_stop(paste(
         "You must {ui_field('add_reviewer')} for a single reviewer at time.\n",
         "Reviewer's name must be a single string."
     ))
 
-    if (!length(get_gold_classes(x))) ui_stop(paste(
-        "{ui_code('x')} has no gold classes set up.\n",
+    if (!length(get_gold_classes(gold))) ui_stop(paste(
+        "{ui_code('gold')} has no gold classes set up.\n",
         "Please, set them by {ui_code('set_gold_classes()')} before to {ui_code('add_reviewer()')}"
     ))
 
-    rev_id <- length(get_reviewers(x)) + 1L
+    has_balance <- length(get_balanced_variables(gold))
+    if (!has_balance) {
+        if (ui_nope(paste0(
+            "This goldrake has not been balanced by any variable.\n",
+            "You cannot balance the selection after adding reviewers.\n",
+            "\n",
+            "Are you sure to continue without any balancing variables?"
+        ))) {
+            ui_fail("Process interrupted. Reviewer was not added.")
+            return(invisible(gold))
+        }
+    }
+
+    rev_id <- length(get_reviewers(gold)) + 1L
     rev_code <- paste0("rev_", rev_id)
     reviewer <- tibble::tibble(
         id   = rev_id,
@@ -61,27 +75,40 @@ add_reviewer.goldrake <- function(
 
     full_name <- stringr::str_c(name, surname, sep = ", ")
 
-    old_reviewers <- get_reviewers(x)
+    old_reviewers <- get_reviewers(gold)
     if (any(full_name == old_reviewers)) ui_stop(paste(
         "Reviewer {ui_value(full_name)} already present.",
         "Please, provide a different one."
     ))
 
-    attr(x, "reviewer_names") <- old_reviewers %>%
+    attr(gold, "reviewer_names") <- old_reviewers %>%
         c(purrr::set_names(full_name, rev_code))
 
-    x[["reviewers"]] <- x[["reviewers"]] %>%
-        dplyr::bind_rows(reviewer)
+    gold[["reviewers"]] <- dplyr::bind_rows(gold[["reviewers"]], reviewer)
 
     set.seed(seed)
-    if (!length(x[["used_data"]])) {
-        x[["used_data"]] <- tibble(
-            sampled_obs = nrow(x[["original_data"]]) %>%
-                sample.int()
+    if (!length(gold[["used_data"]])) {
+
+        base_data <- gold[["original_data"]] %>%
+            dplyr::mutate(.goldid = dplyr::row_number())
+
+        if (has_balance) {
+            base_data <- base_data %>%
+                dplyr::group_by_at(get_balanced_variables(gold))
+        }
+
+        balanced <- base_data %>%
+                dplyr::sample_frac() %>%
+                dplyr::mutate(n = dplyr::row_number()) %>%
+                dplyr::ungroup()
+
+        gold[["used_data"]] <- tibble(
+            sampled_obs = balanced[[".goldid"]][order(balanced[["n"]])]
         )
     }
 
-    x[["used_data"]][[rev_code]] <- factor(NA, levels = get_gold_classes(x))
+    gold[["used_data"]][[rev_code]] <- factor(NA, levels = get_gold_classes(gold))
     ui_done("Reviewer {ui_value(full_name)} added.")
-    invisible(x)
+
+    invisible(gold)
 }