+++ NEWS.md
@@ -2,7 +2,7 @@
 * Added `print()` method for goldrake
 * Added `balanced_group_by()`
 * Added `set_gold_classes()`
-* Added `goldrake()` costructor
+* Added `goldrake()` constructor
 * Added support for tidy evaluation
 * Updated `README`;
 * Added `utils.R`, `directory.R`, and `ui.R` with some feature from